set up travis and appveyor!
- supertuxsmash, along with sqee, now builds without warnings using MSVC
- appveyor now generates a ready to use (debug) windows build
- detection of platform/compiler and finding of sqee fixed in CMakeLists.txt
- also fixed a few runtime errors that msvc picked up, plus a few other small bugs
- moved controller configs to a directory

+++ src/game/FighterPrivate.cpp
@@ -1033,7 +1033,7 @@ void Fighter::base_tick_fighter()
 
     update_animation();
 
-    mArmature.compute_ubo_data(current.pose, mBoneMatrices.data(), mBoneMatrices.size());
+    mArmature.compute_ubo_data(current.pose, mBoneMatrices.data(), uint(mBoneMatrices.size()));
 
     mModelMatrix = maths::transform(Vec3F(current.position, 0.f), current.rotation, Vec3F(1.f));
 }
